+++ src/mlr.jl
@@ -84,8 +84,8 @@ end
     mlrchol(X, Y)
     mlrchol(X, Y, weights::Weight)
     mlrchol!mlrchol!(X::Matrix, Y::Matrix, weights::Weight)
-Compute a mutiple linear regression model (MLR) 
-using the Normal equations and a Choleski factorization.
+Compute a mutiple linear regression model (MLR) using the Normal equations 
+    and a Choleski factorization.
 * `X` : X-data, with nb. columns >= 2 (required by function cholesky).
 * `Y` : Y-data (n, q).
 * `weights` : Weights (n) of the observations.